pyproject.toml
[project]
name = "papra-llm-manager"
version = "0.1.0"
description = "Enhance Papra with AI-powered text extraction and intelligent tagging"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.76.0",
    "click>=8.3.1",
    "httpx>=0.28.1",
    "openai>=2.15.0",
    "pdf2image>=1.17.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[project.scripts]
papra-llm = "papra_llm_manager.cli:main"
